requirements.txt
aiogram==3.4.1
gigachat==0.1.20
python-dotenv==1.0.0
aiofiles==23.1.0
aiohttp==3.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
curl-cffi==0.6.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
propcache==0.1.0
pydantic==1.10.13
requests==2.31.0
urllib3==2.0.4
yarl==1.9.2
httpx==0.27.0